+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.7.1 - Jun 29 2021
+
+- Changed `Request.Params` is be a getter method.
+- Added `Request.WithParams` to set params.
+
 ## v0.7.0 - Jun 28 2021
 
 - Added http.HandlerFunc compatibility mode to help working with `http.HandlerFunc`.